+++ src/experipy/__init__.py
@@ -9,7 +9,7 @@ from . import (
 )
 
 
-__version__     = "0.3.1"
+__version__     = "0.4.0"
 
 __title__       = "experipy"
 __description__ = "A framework for writing and running Computational Science experiments"